requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pytz>=2023.3 